+++ gems_views_builder/__init__.py
@@ -26,6 +26,7 @@ from gems.model.parsing import (  # type: ignore
     VariableSchema,
 )
 
+from gems_views_builder.aggregator import Aggregator
 from gems_views_builder.input.calendar import Calendar, load_calendar
 from gems_views_builder.input.catalog import (
     Catalog,
@@ -35,7 +36,9 @@ from gems_views_builder.input.catalog import (
     TermsOperator,
     TimeOperator,
     load_catalog,
+    load_catalogs,
 )
+from gems_views_builder.input.input_data import InputData
 from gems_views_builder.input.library import Library, load_library
 from gems_views_builder.input.simulation_table import (
     FilteredSimulationTable,
@@ -43,9 +46,10 @@ from gems_views_builder.input.simulation_table import (
     filter_simulation_table,
     load_simulation_table,
 )
-from gems_views_builder.input.system import System
+from gems_views_builder.input.system import System, load_system
 from gems_views_builder.input.taxonomy import Taxonomy, TaxonomyCategory, TaxonomyItem, load_taxonomy
 from gems_views_builder.input.view_config import TimeAggregation, ViewConfig, load_view_config
+from gems_views_builder.loader import Loader
 from gems_views_builder.metric_view import MetricView
 from gems_views_builder.metrics_builder import (
     MetricStructure,
@@ -54,9 +58,10 @@ from gems_views_builder.metrics_builder import (
     build_metric_structure,
 )
 from gems_views_builder.views_builder import ViewBuilder
-from gems_views_builder.writer import MergedView
+from gems_views_builder.writer import MergedView, Writer
 
 __all__ = [
+    "Aggregator",
     "Calendar",
     "load_calendar",
     "FilteredSimulationTable",
@@ -65,6 +70,8 @@ __all__ = [
     "filter_simulation_table",
     "Catalog",
     "load_catalog",
+    "load_catalogs",
+    "InputData",
     "Metric",
     "PropertySchema",
     "Term",
@@ -94,7 +101,10 @@ __all__ = [
     "MetricStructureTable",
     "build_metric_structure",
     "MetricView",
+    "Loader",
     "MergedView",
     "ViewBuilder",
+    "Writer",
     "System",
+    "load_system",
 ]
