+++ tests/test_catalog_taxonomy_validator.py
@@ -14,9 +14,7 @@ from pathlib import Path
 
 import pytest
 
-from gems_views_builder.input.catalog import Catalog, load_catalog, load_catalogs
-from gems_views_builder.input.taxonomy import load_taxonomy
-from gems_views_builder.loader import Loader
+from gems_views_builder import Catalog, Loader, load_catalog, load_catalogs, load_taxonomy, load_view_config
 from gems_views_builder.validation.catalog_taxonomy_validator import (
     validate_catalog_against_taxonomy,
     validate_catalogs_against_taxonomy,
@@ -30,8 +28,6 @@ def test_validate_catalog_against_taxonomy_passes_for_test_dataset(test_dataset_
 
 
 def test_validate_catalogs_against_taxonomy_passes_for_test_dataset(test_dataset_dir: Path) -> None:
-    from gems_views_builder.input.view_config import load_view_config
-
     taxonomy = load_taxonomy(test_dataset_dir / "taxonomy.yml")
     view_config = load_view_config(test_dataset_dir / "view_config.yml")
     catalogs = load_catalogs(test_dataset_dir, view_config.catalog_ids)
