+++ tests/test_model_library.py
@@ -20,7 +20,7 @@ from gems.model.parsing import (  # type: ignore[import-untyped]
     VariableSchema,
 )
 
-from gems_views_builder.input.library import Library, load_library
+from gems_views_builder import Library, load_library
 
 
 def _library_path(test_dataset_dir: Path) -> Path | None:
